Adapt kuttl test for SNO with multiple nodes
Remove checks for the number of pods which depends on the number of
nodes because they are being deployed with a daemonset. Instead,
check that number of OVNController pods is the same as the number of
nodes using a script qand check for successful deployment of the
OVNController by checking the conditions.