Add DNS entry for ovsdbserver- services
When the ovsdbserver-sb/nb pod gets deleted we can't ensure that it will
be recreated with the same internalapi IP, since that IP is popullated
to the EDPM nodes during ansibleee-deployment_phase, if the IP changes
during reboot of the pod EDPM won't know until user retriggers
ansibleee-deployment_phase. This will mean that meanwhile ovn_controller
and neutron-ovn-metadata won't have connectivity to the SB DB.

In order to fix this instead of using a string of IPs on the ovn-remote
a single DNS entry will be used. Every service will add an entries to
the openstack-dnsmasq:
 - ovsdbserver-xb.openstack.svc

The last one will be the one popullated to the EDPM node, as querying
it will return one IP from all the SB pods initialized at that moment
(dns will use sequential round-robin to fulfill the request).

Depends-on: https://github.com/openstack-k8s-operators/install_yamls/pull/686
Resolves: OSPRH-660
(cherry picked from commit 5882e62f2487e282921c567e92d348fc8a79e738)