[EnvTest] Fix false positives checks for Config Maps
The tests were passing as ListConfigMaps expected
namespace as parameter but a non-existing namespace was
provided as input.
Also drop the tests which checks config map removals
on CR deletion, as the cleanup of these built-in resources
is done by k8s controller once parent resource is deleted
but that doesn't run in EnvTest so such cleanup never happens.
Instead we can just check if OwnerReferences are set correctly
on the created Config Maps.

Noticed this while working on pr #153.