Do not rely on NodeName for ovn-config job
There is a max limit of 63 characters on the job name,
and using Instance Name + NodeName(with fqdn) can surpass
this limit and cause config jobs to not be created.

Instead switch the job names to rely on ovn-controller
pod names which will already be unique and size
length is constant as pod names are based on daemonset Name.

Also when External Endpoint is not set[1] we don't need
to return as it will never be set when DBClusters are running
without NetworkAttachment. For NetworkAttachment case it's
already taken care as part of generateExternalConfigMaps.

Added functional tests for the same.

[1] https://github.com/openstack-k8s-operators/ovn-operator/pull/153
Resolves: https://issues.redhat.com/browse/OSPRH-683