Disable logging to file (mostly)
The service still logs a single line into the file before seizing. This
is because we pass --log-file to it. And we can't avoid it because
ovs-ctl will always pass the (default, inaccessible) log-file argument
to the service if no other file is passed.

(Ideally, ovs-ctl would be patched to allow to avoid passing any
--log-file to the service. But this is beyond the scope of this work.)