ovndbcluster: get rid of postStart
This patch moves the code from settings.sh into setup.sh.

The settings will still happen, but they will not block transition of
the pod into Ready state, which is important to do even if raft cluster
is not re-built, yet. This is in line with k8s guidelines about not
making liveness / readiness checks depend on other pod states (in this
case - on the connectivity to other raft members.)