Increase termination grace period for raft members
It's important to give them a good chance to communicate to the leader
that they are leaving. If the leader is offline for a bit (pod deleted),
the message may not have enough time to propagate and commit into the
leader database (the default is 30 seconds), which may leave the cluster
in divergent state.