Remove the double update when finalizer added
The northd and dbcluster controller did a separate client.Update after
it added the self finalizer then returned and run the defer call that
did another client.Patch. However the first client.Update only wrote the
main resource and ignored the status subresource. However it refreshed
the whole instance with the result of the Update meaning it reset the in
memory status updates. We don't need two separate update so this PR
removes the standalone Update and rely on the deferred Patch call to
do all the update.