Move helper creation right after the instance is read
The helper creation saves a copy of the instance to use it at the end of
the reconcile to generate the diff for the client.Patch call. Therefore
any changes to the instance before the helper created is lost from the
diff perspective and not persisted. Both northd and dbcluster had the
instance status init before the helper creation and when the
observedGeneration bump is moved there too the observedGeneration
changes started to be ignored as it was not visible in the diff the
helper see.

This patch moves the helper creation up in the reconcile loop.