Fix default handling for LogLevel param
When OVNDBCluster or OVNNorthd CR is created with empty spec
with golang structs(as done in functional tests or in openstack-operator)
the golang defaults get's applied as per type, in this
case logLevel being string type an empty string("") get's set and
that leads to kubebuilder default("info") not honored.

[1] provide more details on this behavior as kubebuilder
default is different here then the golang default for the string type.
Handle it by adding omitempty field, with this "" as empty value
will be discarded and kubebuilder defaults will work.

[1] https://github.com/gibizer/operator-lint/tree/main/linters/crd/C003
Closes: OSPRH-6597