Remove non graceful OVN DB pod delete Kuttl test
We are seeing a failure with the `06-delete-pods-non-graceful.yaml` job
in OCP 4.15 Prow CI. After discussing with the ovn-operator maintainers
we decided this test is invalid. With both `--force` and
`--grace-period=0` the database lock file is not being released and this is something that requires hands on recovery which isn't being covered in this test case.

JIRA: https://issues.redhat.com/browse/OSPRH-7515