Fix gateway datapath disrupt on update
Reimplementing the vswitchd reload using separate start and stop scripts
so that it can be executed partially between consecutives instances of
the pod. To avoid disruptions, it runs ovs save-flows command on the
stop-vswitchd script, saving the resulting backup on a mounted folder
from the host and then loading that backup on the start script. For this
to succeed, the script needs to set the flow-restore-wait flag from ovs
to true while loading the flows.

Note that ovs save-flows command needs to have ovsdb-server running, so
a semaphore like mechanism was added to ensure the ovsdb-server
container is never deleted before ovs-vswitchd.

Closes-Issue: OSPRH-6326