Move db initialization to an init container
There is a potential race condition where ovs-vswitchd containers
can run ovs db commands and fails while ovsdb-server stops and restarts
after db initialization. This makes ovs-vswitchd container to fail
and restart and it eventually recovers and it's better to avoid this.

This patch moves the db initialization part to init container to
avoid this race.

Resolves: OSPRH-637