Generate External config map on OVNDBCluster
Due to openstack-operator PR#1104 deployment can run without
OVNController, currently it is in charge of create the
External ConfigMap used by EDPM deployment.

Moving the creation of the map on OVNDBCluster to ensure the
map exist regardless of the setup of the enviroment.

The map's name is hardcoded (ovncontroller-config) this is done
on purpose, as this name is hardcoded on the openstack-operators
config files. This should be addressed on the future, but for now
we'll handle it this way.

With this commit OVNDBCluster will start watching OVNController CR
to ensure that any change made in
OVNController.Spec.ExternalIDs.OvnEncapType is monitored and can
update the configMap accordingly.

Resolves: OSPRH-10372