Allow exposing ovn DB using a k8s service
Using multus to attach the ovn db pods to the networks
that the edpm nodes can access it won't work for BGP. A
loadbalancer service, like metallb, could be used in this case.

This introduces service override to the ovndbcluster, like we
have for other services which use a LB service to expose them
to the outside. Adding the AnnotationHostnameKey annotation
the service hostname gets automatic registered in the DNS instance.

In general this could replace all the current implementation of
attaching the ovn db pods using multus, creating the headless
service and creating the dnsdata to register it in dns. For
update/backward compatability this is kept as if if no service
override is provided.

To expose the ovndbcluster-sb via a metallb loadbalancer would
look like:

```
  ovn:
    enabled: true
    template:
      ...
      ovnDBCluster:
        ovndbcluster-nb:
	  ...
        ovndbcluster-sb:
          dbType: SB
	  ...
          override:
            service:
              metadata:
                annotations:
                  metallb.universe.tf/address-pool: internalapi
                  metallb.universe.tf/allow-shared-ip: internalapi
                  metallb.universe.tf/loadBalancerIPs: 172.17.0.80
              spec:
                type: LoadBalancer
      ovnNorthd:
        ...
```

Jira: OSPRH-651

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>