Allow ovs-vswitchd logging by separating restore flows
We used --detach on ovs-vswitchd execution to allow for flow
restoration, but this had the side effect of not having its console logs
available. In order to get those back, we use --log-file and then tail
that file in the end of the script to be able to see these logs from the
oc log command.

Jira: OSPRH-11228