[ovn-controller-ovs] Compact OVSDB in init container
If OVSDB get's large enough it impacts start up time.
It is observed in a case for a DB ~ 2.5 GB it took
60+ seconds to start ovsdb which is higher than
configured liveness probes and due to this pod got
stuck into a restart loop.
With this patch compacting OVSDB in init container
if the db exists to avoid this issue.

Resolves: OSPRH-17558