Change liveness and readiness probes for the ovn-northd POD
Before this patch liveness and readiness probes for the ovn-northd POD
were just checking PID of the ovn-northd process.

This patch changes it to use custom script which checks status of the
connection of ovn-northd container to the NB and SB databases.

The reason why that was choosen instead of simply checking ovn-northd
status is the fact that status shows `standby` in case if ovn-northd is
actually not connected to the DB but also if there is more than one
ovn-northd PODs, only one can be active and all others are always in
`standby` which is completely normal thing.

Closes: #OSPRH-8042
Signed-off-by: Slawek Kaplonski <skaplons@redhat.com>