Accept IPv4 and IPv6 connections on ovsdbserver
Until now we checked if the environment is using IPv4 or IPv6
in /etc/hosts, which depends on how the OCP control plane is specified.
This is not accurate, since we can have IPv4 on the OCP control plane
but IPv6 on the OCP dataplane. Therefore we implemented [::] as
listening entry for both IPv4 and IPv6 environments. Even though we are
not currently expecting dual-stack configurations [0] this works
because OVN does have dual-stack capabilities.

[0] https://github.com/openstack-k8s-operators/install_yamls/blob/main/scripts/gen-netatt.sh#L40

Resolves: OSPRH-5808