Do not enable Northd Metrics when exporterImage is unset
For minor update scenarios we need to honor exporterImage as
in older versions will not have OpenstackNetworkExporterImage
set so exporterImage="".
So with this patch metrics will be enabled only when
metricsEnabled=true and exporterImage is set.

Also removed the default image as now included in
openstack-operator like other images, this was temporary
change.

Also dropped unnecessary log message when deployment do not exist.

Related-Issue: https://issues.redhat.com/browse/OSPRH-12568
Depends-On: https://github.com/openstack-k8s-operators/openstack-operator/pull/1587