Set logging level in ovn-controller
OVNLogLevel affects ovncontroller logs and OVSLogLevel affects
ovs-vswitchd and ovsdb-server. Both OVNLogLevel and OVSLogLevel can take
values from the vlog range: off, emer, err, warn, info, or dbg See
ovs-appctl(8) for a definition of each log level. The change of the log
level will be modified from the config pod and not on the pod itself to
avoid restarting ovn-controller and ovn-controller-ovs pods.

Resolves: OSPRH-6429
Signed-off-by: Elvira Garcia <egarciar@redhat.com>