Change retention policy of the OVNDBCluster PVCs to be "retained"
Synchronization of the OVN dbs with Neutron is pretty heavy and manual
operation it is benefical to not delete those OVN databases when
OVNDBCluster's PODs are deleted.
To achieve that this patch removes `OwnerReference` attribute from the
PVCs created for the OVNDBClusters so that they are not cleaned by the
OpenShift's garbage collector and it also changes
PersistentVolumeClaimRetentionPolicy when deleted and when scaled to the
same `Retain` policy.

Depends-On: https://github.com/openshift/release/pull/71153

Closes: #OSPRH-19930

Signed-off-by: Slawek Kaplonski <skaplons@redhat.com>